control: extract ControlConnection to a module of its own
